+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ## 6.1.0 (upcoming)
 
+- Download Java 11 if only Java 8 is installed (see [#39](https://github.com/valentjn/vscode-ltex/issues/39) and [documentation](https://valentjn.github.io/vscode-ltex/docs/deprecation-of-java-8.html))
 - Upgrade bundled AdoptOpenJDK JRE to 11.0.8+10
 - Update some NPM dependencies